Build against glibc 2.26
Previously we were building against glibc 2.26.

amazonlinux 2 has 2.26, and we have users wanting to use this, so let's
jump back a couple of glibcs to support them.

Another advantage here is that amazonlinux maintains `yum` working out
of the box, rather than needing to configure mirrors.

A down-side is that the `node` version used by the `checkout` action
links a newer glibc than 2.26. For now we activate a hatch to revert to
an older node, but in the future we may need to work around another way,
e.g. by manually checking out the source ourselves instead of relying on
the actions action which relies on node. We also now perform container
set-up before the checkout action, so we can set up the tools it assumes
(e.g. `tar`)