Build toolchain with musl on Linux (#46)
On Linux, this change introduces a two-stage bootstrap process in which
we first build a musl toolchain targeting the host and then use it to
build a toolchain that builds with musl for the target platform but is
itself statically linked against musl.

This avoids any glibc dependency, which allows us to remove the custom
containers and instead rely on the latest Ubuntu runners. ARM runners
are now in public preview and replace buildjet.